Use port numbers instead of names in prometheus annotations.
Prometheus scrape configs often use the `prometheus.io/port` annotation
to configure targets per pod:

```
- source_labels: [__address__, __meta_kubernetes_pod_annotation_prometheus_io_port]
  action: replace
  regex: ([^:]+)(?::\d+)?;(\d+)
  replacement: $1:$2
  target_label: __address__
```

This rule only matches when the `prometheus.io/port` annotation is a
port number, and it's not clear to me how the scrape config could be
modified to work with port names. To make the chart compatible with
standard prometheus scrape configs, this patch changes the default port
annotation from port names to port numbers.

Signed-off-by: Josh Carp <jm.carp@gmail.com>