feat(nginx): ensure dns resolver respects a ttl
problem:
when services behind the nginx proxy roll or scale up, nginx does not pick
up the new ips as it never rechecks DNS until it is restarted. this can
be mitigated by setting a ttl with the valid parameter on the DNS
resolver AND by ensuring the domain lookups are made through a variable.

see: https://serverfault.com/questions/240476/how-to-force-nginx-to-resolve-dns-of-a-dynamic-hostname-everytime-when-doing-p/593003#593003

solution:
ensure the resolver has a suitable validity period and move domains
behind an nginx variable, forcing nginx to recheck DNS when proxying
requests.

Signed-off-by: Cassandra Comar <cassandra@ditto.live>